+++ R/StatecCMP/R/zzz.R
@@ -1,4 +1,4 @@
 .onAttach <- function(libname, pkgname) {
-	packageStartupMessage("Statec functions and convenience wrappers, version 0.0.5, 2023-02-16.")
+	packageStartupMessage("Statec functions and convenience wrappers, version 0.0.6, 2023-02-22.")
   packageStartupMessage("Maintained by Michel Geller (michel.geller@statec.etat.lu),\npreviously by Andrei V. Kostyrka (andrei.kostyrka@gmail.com).")
 }